% Generated by roxygen2 (4.0.0): do not edit by hand
\name{registry}
\alias{registry}
\title{Some file in which to store temporary information used by the package
(such as which questions have already been seen).}
\usage{
registry()
}
\description{
Some file in which to store temporary information used by the package
(such as which questions have already been seen).
}

