% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{socompanion}
\alias{socompanion}
\alias{socompanion-package}
\title{StackOverflow R Companion provides alerts from within the interactive R
console whenever a new R question is posted. Since useRs tend to spend a lot of time
in the console, by embedding the read-answer flow into the console it should greatly
increase the speed at which questions can be answered -- which leads to more of that
juicy karma!}
\description{
StackOverflow R Companion provides alerts from within the interactive R
console whenever a new R question is posted. Since useRs tend to spend a lot of time
in the console, by embedding the read-answer flow into the console it should greatly
increase the speed at which questions can be answered -- which leads to more of that
juicy karma!
}

